BT Real Estate Project
Django real estate website created after courses This is my first real web application. I coped with many hard tasks, one of which was emergency migration from PyCharm to VSCode because of functionality lack in PyCharm Community. Another thing was dealing with versions of Django. At first I was using 3.0 but there are functions that are not compatible with python 3.6, so I migrated to Django 2.1 and it was ok. I did everything myself except Bootstrap theme. You can find my instructor in Github - https://github.com/bradtraversy, and even find the same project in his repos but mine has some distinctions:
1) code is PEP formatted
2) fixed some minor bugs in text
3) filled links in navbars
4) inserted real text in some places (I just wanted it to look more natural)
Overall, I love this project. This wasn't easy for novice as me but it was very interesting.